[add]キャラうごいた

+++ Assets/Scripts/src/positions/Position2.cs
@@ -2,13 +2,13 @@ namespace src.positions
 {
     public class Position2
     {
-        public int X, Y;
+        public float X, Y;
 
         public Position2() : this(0, 0)
         {
         }
         
-        public Position2(int x, int y)
+        public Position2(float x, float y)
         {
             X = x;
             Y = y;
